services/ml-models/derm-foundation/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
transformers==4.36.0
torch==2.1.0
pillow==10.1.0
numpy==1.26.2
scikit-learn==1.3.2
opencv-python-headless==4.8.1.78
